add elasticsearch

+++ mall-system/src/main/java/me/shaw/mall/service/OrderService.java
@@ -4,7 +4,7 @@ import me.shaw.mall.model.Order;
 
 public interface OrderService {
 
-    Order createOrder(float transaction, String status);
+    Order createOrder(int transaction, String status);
 
     void cancelOrder(Long OrderId);
 }
